Refactor: move files to correct Persistence subfolders, remove cos listing, split Game constructor, delete analysis doc
Agent-Logs-Url: https://github.com/ChrisWiGit/CivOne/sessions/ac9a1ae8-2ba0-417d-bec4-d0d1a6e62bb3
