+++ src/Persistence/Model/GameStateDtoMapper.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using CivOne.Persistence.Stubs;
 using CivOne.Tiles;
 using CivOne.Units;
