--- src/Persistence/Model/ITerrainFactory.cs
@@ -1,13 +1,0 @@
-using CivOne.Enums;
-using CivOne.Tiles;
-
-namespace CivOne.Persistence.Model
-{
-	/// <summary>
-	/// Factory for creating <see cref="ITile"/> instances from terrain type and coordinates.
-	/// </summary>
-	public interface ITerrainFactory
-	{
-		ITile CreateTile(Terrain terrain, int x, int y, bool special);
-	}
-}
